[#72]; feat: 필드 파싱 교필 조회 API 적용

+++ src/main/kotlin/com/yourssu/soongpt/domain/course/application/dto/MajorRequiredRequest.kt
@@ -6,7 +6,7 @@ import org.hibernate.validator.constraints.Range
 
 data class MajorRequiredRequest(
     @Range(min = 15, max = 25, message = "학번은 15부터 25까지 가능합니다.")
-    val schoolId: Long,
+    val schoolId: Int,
 
     @NotBlank
     val department: String,
